refactor: prediction mode

+++ ultralytics/tracknet/dataset.py
@@ -145,7 +145,7 @@ class TrackNetDataset(Dataset):
 
         img = self.get_image_cache(d['cache_npy'])
 
-        img = torch.from_numpy(img)
+        img = torch.from_numpy(img).float()
         target = torch.from_numpy(d['target'])
 
         img_files = [f"{self.root_dir}/../{im}" for im in d['img_files']]
